Add a language plugin.
This plugin allows to link pages that are translated to another
language. Each resource should have a `language` meta attribute and a
`uuid` meta attribute. Resource that have those two meta attributes
will get a new attribute with the list of other resources in a
different language that have the same `uuid` meta attribute.

It is then possible to use something like this to grab the list of
available languages for a given resource:

{% if resource.translations is defined %}
<p>Other languages</p>
<ul>
  {% for translation in resource.translations %}
    <li><a href="{{ translation.full_url }}">
       {{ languages[translation.meta.language] }}
        </a></li>
  {% endfor %}
</ul>
{% endif %}

(You need a `languages` context in `site.yaml` for this snippet to work:
    context:
      data:
        languages:
          fr: Français
          en: English
)