Added -o/--override option to the gen subcommand to set config properties at runtime.
This is most useful in the form of "-o mode=production", so you can automate generating a production version of the site without messing with the config files.

But you can also use it to set more involved metadata, such as "-o context.data.menu#0.description=Whatever"

Note that for parsing simplicity, you index an array as menu#0 instead of menu[0].

There's probably a better way to go about all this, but!