Git Plugin: Fall back on file mtime/ctime if not tracked by git
When using 'git' as the value for resource.meta.created or
resource.meta.modified, the Git plugin will fetch those from the git
repository. But if the file was not known by git, it used to no alter
the attributes, leaving them as the 'git' string.

This was causing issues because many consumers of those attributes
(for instance, the xmldatetime template filter from jinja2) expect to
find a datetime instance and will raise an exception if they find a
string.

This patch works around this issue by falling back on the file's own
ctime and mtime in case the file wasn't added to the git repository yet
(which can happen for brand new entries).