Fix generation of atom.xml in the basic layout.
When using 'resource.url' as the argument for the {% refer %} jinja2
tag, hyde fails to find the resource because since the url starts with a
'/', it is considered as an absolute path by the hyde FS abstraction.
This results in an empty content in the atom feed.

Using resource.relative_deploy_path solves the issue, as it is basically
the same as the url without the initial '/'.