Fixed bug where syntax highlighted code was being run through markdown
Added leading new line to HTML before running through syntax highlighting. Without this the div block can on the next line to normal text, which means that it isn't treated as a HTML block by the markdown parser, which means the markdown parser will parse the contents of the <span> tags and corrupt the code. This means that * symbols are replaced with em tags in code.