[WIP] Add Sphinx docs
This is the starting point of documenting Hyde using Sphinx. This
provides several benefits:

1. Code and documentation live together, so it's easier to track and
enforce that there exists documentation for new or changing code.

2. This lets us get Hyde on Read The Docs.

3. We can use things like `autoclass` where appropriate to pull
documentation directly from source code without reinventing the wheel.

I've tried to replicate what currently exists on hyde.github.io. The
first pass of this effort is concerned with parity, and we can continue
to improve from there on. Steps:

1. Convert all existing documentation on hyde.github.io to Sphinx. This
is covered by this commit and should be considered "in review".

2. Open issues for core parts of Hyde that are missing documentation.

3. Close these issues individually. This is a great opportunity for
community members wishing to get involved to contribute.

4. Require significant changes in code to be accompanied by updated /
new documentation the same way we would require tests.

While these steps are in progress, the following should also be done:

1. Deprecate the docs at hyde.github.io, linking users to Read The Docs.

2. Replace the content currently hosted at hyde.github.io with a new
brochure site showing off some of Hyde's uses and linking to other
interesting content (e.g. a list of themes, links to sites powered by
Hyde, etc.).

References #188, #209, and quite a few other issues.

@lakshmivyas, @llonchj: please review. This is a starting point for
Sphinx docs, and the PR is labelled as a work in progress, but I have no
specific threshold for it being "done". Once we're all satisfied with
the first pass, I'll set up RTD, and we can start filing issues for
specific areas lacking docs.